requirements.txt
aiohttp==3.9.3
aiohttp-socks==0.8.4
aioimaplib==1.0.1
aiosignal==1.3.1
aiosmtplib==3.0.1
antlr4-python3-runtime==4.9.3
anyio==4.2.0
appdirs==1.4.4
appnope==0.1.3
argon2-cffi==23.1.0
argon2-cffi-bindings==21.2.0
arrow==1.3.0
astroid==3.0.2
asttokens==2.4.1
async-lru==2.0.4
async-timeout==4.0.3
attrs==23.2.0
Babel==2.14.0
base58==2.1.1
beautifulsoup4==4.12.3
bech32==1.2.0
bleach==6.1.0
cachetools==5.3.2
cattrs==23.2.3
cbor2==5.5.1
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
colorama==0.4.6
comm==0.2.1
contourpy==1.2.0
cpe==1.2.1
cryptography==41.0.7
cycler==0.12.1
debugpy==1.8.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
exceptiongroup==1.2.0
executing==2.0.1
fastjsonschema==2.19.1
fonttools==4.47.2
fqdn==1.5.1
frozenlist==1.4.1
google==3.0.0
google-api-core==2.16.2
google-auth==2.27.0
google-cloud-bigquery==3.17.1
google-cloud-core==2.4.1
google-cloud-vision==3.6.0
google-crc32c==1.5.0
google-resumable-media==2.7.0
googleapis-common-protos==1.62.0
grpcio==1.60.1
grpcio-status==1.60.1
html5lib==1.1
idna==3.6
importlib-metadata==7.0.1
importlib-resources==6.1.1
iniconfig==2.0.0
ipykernel==6.29.0
ipython==8.18.1
ipywidgets==8.1.1
isoduration==20.11.0
isort==5.13.2
jedi==0.19.1
Jinja2==3.1.3
json5==0.9.14
jsonpointer==2.4
jsonschema==4.17.3
jsonschema-specifications==2023.12.1
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter-lsp==2.2.2
jupyter_client==8.6.0
jupyter_core==5.7.1
jupyter_server==2.10.0
jupyter_server_terminals==0.5.2
jupyterlab==4.0.12
jupyterlab-widgets==3.0.9
jupyterlab_pygments==0.3.0
jupyterlab_server==2.24.0
kiwisolver==1.4.5
lark==1.1.9
lmdb==1.4.1
MarkupSafe==2.1.4
matplotlib==3.8.2
matplotlib-inline==0.1.6
mccabe==0.7.0
mistune==3.0.2
more-itertools==10.2.0
msgpack==1.0.7
multidict==6.0.4
nbclient==0.9.0
nbconvert==7.14.2
nbformat==5.9.2
nest-asyncio==1.6.0
notebook==7.0.7
notebook_shim==0.2.3
numpy==1.26.3
overrides==7.7.0
packaging==23.2
pandas==2.2.0
pandocfilters==1.5.1
parso==0.8.3
pexpect==4.9.0
pillow==10.2.0
platformdirs==4.1.0
pluggy==1.4.0
prometheus-client==0.19.0
prompt-toolkit==3.0.43
proto-plus==1.23.0
protobuf==4.25.2
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.7
pyarrow==15.0.0
pyasn1==0.5.1
pyasn1-modules==0.3.0
pycparser==2.21
pycryptodome==3.19.1
Pygments==2.17.2
pylint==3.0.3
pyOpenSSL==23.2.0
pyparsing==3.1.1
pyrsistent==0.20.0
pyspark==3.5.0
pytest==8.0.0
python-bitcoinlib==0.12.2
python-dateutil==2.8.2
python-dotenv==1.0.1
python-json-logger==2.0.7
python-socks==2.4.4
pytz==2024.1
PyYAML==6.0.1
pyzmq==25.1.2
qtconsole==5.5.1
QtPy==2.4.1
referencing==0.33.0
regex==2023.12.25
requests==2.31.0
requests-cache==1.1.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rpds-py==0.17.1
rsa==4.9
scalecodec==1.2.7
seaborn==0.13.2
Send2Trash==1.8.2
simplejson==3.19.2
six==1.16.0
sniffio==1.3.0
soupsieve==2.5
stack-data==0.6.3
stix2-patterns==2.0.0
stix2-validator==3.1.4
terminado==0.18.0
tinycss2==1.2.1
tomli==2.0.1
tomlkit==0.12.3
tornado==6.4
traitlets==5.14.1
types-python-dateutil==2.8.19.20240106
typing_extensions==4.9.0
tzdata==2023.4
uri-template==1.3.0
url-normalize==1.4.3
urllib3==1.26.18
vcrpy==5.1.0
wcwidth==0.2.13
webcolors==1.13
webencodings==0.5.1
websocket-client==1.7.0
widgetsnbextension==4.0.9
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
zipp==3.17.0
